Backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
chardet==5.2.0
cssselect2==0.7.0
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==64.0.0
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fonttools==4.54.1
html5lib==1.1
optional-django==0.1.0
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.16.0
reportlab==4.2.5
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
tinycss2==1.3.0
tzdata==2024.1
webencodings==0.5.1
wheel==0.44.0
zopfli==0.2.3
drf-yasg==1.21.8